# src/docs/swagger/collections.yaml
openapi: 3.0.0
info:
  title: RBAC
  description: The API collection for a Role-based access control system
  version: 1.0.0
servers:
  - url: http://localhost:4000/api
    description: RBAC Development Server
  - url: http://api.rbac.shubhampurwar.in/api
    description: RBAC Production Server
...
paths:
...
  /roles:
    get:
      tags:
        - Roles
      description: Allows authenticated users to retrieve a paginated list of roles
      operationId: getRoles
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: A JSON object containing a list of roles
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        title:
                          type: string
                        userCount:
                          type: integer
                          format: int32
                        isActive:
                          type: boolean
                        createdAt:
                          type: string
                          format: date-time
          headers:
            X-Total-Count:
              description: Total number of roles
              schema:
                type: integer
                format: int32

    post:
      tags:
        - Roles
      description: Allows authorized users to add a new role
      operationId: addNewRole
      requestBody:
        description: A JSON object containing role info
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleBody'
      responses:
        '201':
          description: A JSON object containing a new role
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Role'
  /roles/{roleId}:
    get:
      tags:
        - Roles
      description: Allows authorized users to retrieve a role by ID
      operationId: getRoleById
      parameters:
        - in: path
          name: roleId
          description: ID of the role to be retrieved
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A JSON object containing a role
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Role'

    put:
      tags:
        - Roles
      description: Allows authorized users to update an existing role
      operationId: updateRole
      parameters:
        - in: path
          name: roleId
          description: ID of the role to be updated
          required: true
          schema:
            type: string
      requestBody:
        description: A JSON object containing role updates
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleBody'
      responses:
        '200':
          description: A JSON object containing an updated role
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Role'

    delete:
      tags:
        - Roles
      description: Allows authorized users to delete a role
      operationId: deleteRole
      parameters:
        - in: path
          name: roleId
          description: ID of the role to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A JSON object containing ID of the deleted role
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: string
  /roles/{roleId}/activate:
    put:
      tags:
        - Roles
      description: Allows authorized users to activate an inactive role
      operationId: activateRole
      parameters:
        - in: path
          name: roleId
          description: ID of the role to be activated
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A JSON object containing an active role
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Role'
  /roles/{roleId}/deactivate:
    put:
      tags:
        - Roles
      description: Allows authorized users to deactivate an active role
      operationId: deactivateRole
      parameters:
        - in: path
          name: roleId
          description: ID of the role to be deactivated
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A JSON object containing an inactive role
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Role'
components:
  schemas:
...
    Role:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        permissions:
          $ref: '#/components/schemas/RolePermissions'
        userCount:
          type: integer
          format: int32
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RoleBody:
      type: object
      properties:
        title:
          type: string
        permissions:
          $ref: '#/components/schemas/RolePermissions'
    RolePermissions:
      type: array
      items:
        type: string
        enum:
          - view_user
          - activate_user
          - deactivate_user
          - archive_user
          - restore_user
          - delete_user
          - view_role
          - add_role
          - update_role
          - delete_role
          - assign_role
          - unassign_role
          - activate_role
          - deactivate_role
  parameters:
    pageParam:
      in: query
      name: page
      description: Page number for paginated results
      schema:
        type: integer
        format: int32
        min: 1
    limitParam:
      in: query
      name: limit
      description: Number of items per page
      schema:
        type: integer
        format: int32
        min: 1
        max: 100